user points remove after purchase

+++ BattleShipV3/Shared/Data/Commands/User/Update/UpdateUserCommand.cs
@@ -1,4 +1,4 @@
 ﻿
 namespace BattleShipV3.Shared.Data.Commands.User.Update;
 
-public record UpdateUserCommand(string? Name, string? Email, string? Password,int? Elo, int? Points);
+public record UpdateUserCommand(string? Name, string? Email, string? Password, double? Elo, int? Points);
